chore(odyssey-storybook): rearrange files post rebase
