+++ packages/odyssey-storybook/.storybook/preview-head.html
@@ -11,3 +11,14 @@
   href="https://fonts.googleapis.com/css2?family=Inconsolata&display=swap"
   rel="stylesheet"
 />
+
+<style>
+  .action-list-container {
+    max-width: 380px;
+  }
+
+  .action-card-sub {
+    color: $text-sub;
+    font-size: $size-body-caption;
+  }
+</style>
